Add files via upload
Election Audit

Overview of the Election Audit

The purpose of this audit is to provide confirmation results of the election in question. The Colorado Board of Elections have mandated this audit. In this audit there will be several tabulations. Among the tabulations that will be paid special attention to are: Total number of votes casted, votes by candidate, percentage of votes that each candidate won, and the final metric of who won the election in question. 

Election Audit Results:

•	Total votes cast in this congressional election: 369,711
•	County votes:
o	Jefferson: 3,8855 votes, 10.51%
o	Denver: 306055 votes, 82.78% 
o	Arapahoe: 24,801 votes, 6.71%
•	Largest County by vote: Denver
•	Votes by candidate:
o	Charles Casper Stockham: 85,213 votes (23.0%)
o	Diana DeGette: 272,892 votes (73.8%)
o	Raymon Anthony Doane: 11,606 votes (3.1%)

•	Election decision:
o	Winner - Diana DeGette272,892 votes (73.8%)

Election Audit Summary:

The Colorado Board of Elections requisitioned this audit to confirm the popular vote in this Congressional election for the State of Colorado. The rationale behind this exercise is to determine several metrics:

1.	To witness the amount of the population that actively votes within the State of Colorado.
2.	Determine which counties have the highest voting density.
3.	To acquire metadata on voting trends within the State of Colorado.
4.	To determine decisively who is the winner of this Congressional election.

The solution to the aforementioned issues is provided in the Election Audit code that I have submitted with this package. Specifically, this code can be modified to increase the number of counties that participate in any election going forward. This can be done by modifying the counties directory within the code to include more than three counties. The second way that the code can be modified is that it can be used as a control measure to assure that the relevant parties receive the output information (election results) via text file in an untampered manner. The raw data and subsequent results can also be input into a machine learning algorithm that might be able to create predictive models for voting trends. If this is the case, the metadata can be put into simulations with inputs to determine viable strategies for any political candidate. 

Pricing and further consultations are to be negotiated from this point going forward.